NDRS-833: Move incoming signature validation out of component-level logic.
